requirements.txt
aiohttp
aiosignal
async-timeout
attrs
beautifulsoup4
certifi
cffi
chardet
charset-normalizer
discord
discord.py
ffmpeg
frozenlist
idna
multidict
pycparser
PyNaCl
python-dotenv
requests
six
soupsieve
typing_extensions
urllib3
yarl
youtube-dl
